+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Disable assistance button when not in collect mode or when assistance url is not provided.
+
 ## [2.8.0](https://github.com/InseeFr/stromae-dsfr/releases/tag/2.8.0) - 2026-05-20
 
 ### Changed